enhance: wolfram toolkit PR2387
